Add context getter/setter methods only if not already added
Class `FieldBackedProvider` uses ByteBuddy to add a field and its respective getters and setters
to the provided classes. Assuming that we have a class `A` that implements runnable and that we wrap with a `FieldBackedProvider`
`Runnable` interfaces, if later on we use a cglib's `Enancher` class on `A` then the our mechanism will kick in again and try
to add the methods again. This causes a `java.lang.ClassFormatError: Duplicate method name "get__datadogContext$java$lang$Runnable" with signature "()Ljava.lang.Object;"`
to be thrown.

With thic commit we now check that method were not previously defined before adding them, and if they were then we avoid adding them
again.